move quick start page to pages
