+++ src/renderer/pages/quick_start/load.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+export function lazyLoadQuickStartPage() {
+  const LazyPage = React.lazy(async () => {
+    const { createQuickStartPage } = await import("./create");
+    const { Page } = createQuickStartPage();
+    return { default: Page };
+  });
+
+  const Page: React.ComponentType = (props) => (
+    <React.Suspense fallback={null}>
+      <LazyPage {...props} />
+    </React.Suspense>
+  );
+
+  return { Page };
+}
