+++ src/renderer/pages/quick_start/quick_start_header/quick_start_header.tsx
