+++ src/renderer/pages/quick_start/steps/verify_email_step.tsx
@@ -12,7 +12,7 @@ import { useAccount } from "@/lib/hooks/useAccount";
 import { useToasts } from "@/lib/hooks/useToasts";
 import { useServices } from "@/services";
 
-import { QuickStartHeader } from "./QuickStartHeader";
+import { QuickStartHeader } from "../quick_start_header/quick_start_header";
 
 const Container = styled.div`
   margin: 0 auto;
