--- src/renderer/views/LandingView.tsx
@@ -1,22 +1,0 @@
-import Box from "@mui/material/Box";
-import React from "react";
-
-import { QuickStart } from "@/containers/QuickStart";
-import { useQuickStart } from "@/lib/hooks/useQuickStart";
-import { withSlippiBackground } from "@/styles/withSlippiBackground";
-
-const isDevelopment = window.electron.bootstrap.isDevelopment;
-
-export const LandingView = React.memo(() => {
-  const { allSteps, currentStep, nextStep, prevStep } = useQuickStart();
-  return (
-    <Box css={withSlippiBackground} display="flex" style={{ height: "100%", width: "100%" }}>
-      <QuickStart
-        allSteps={allSteps}
-        currentStep={currentStep}
-        onNext={isDevelopment ? nextStep : undefined}
-        onPrev={isDevelopment ? prevStep : undefined}
-      />
-    </Box>
-  );
-});
